Rename/fe/cookie: cookie폴더 util폴더로 이동
